+++ src/components/point-trip.js
@@ -1,4 +1,5 @@
-import {createElement} from "../utils";
+
+import {AbstractComponent} from "./abstract-component.js";
 
 const createPointTrip = (typePoint, city, timeStart, timeEnd, price, orderUber) => {
 
@@ -50,26 +51,18 @@ const createSitePointTripTemplate = (points) => {
   );
 };
 
-export class Point {
+export class Point extends AbstractComponent {
   constructor(points) {
-    this._points = points;
+    super();
 
-    this._element = null;
+    this._points = points;
   }
 
   getTemplate() {
     return createSitePointTripTemplate(this._points);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
   }
 }
