+++ src/components/site-menu.js
@@ -1,4 +1,5 @@
-import {createElement} from "../utils";
+
+import {AbstractComponent} from "./abstract-component.js";
 
 const createSiteMenuTemplate = () => {
   return (
@@ -9,24 +10,8 @@ const createSiteMenuTemplate = () => {
   );
 };
 
-export class Menu {
-  constructor() {
-    this._element = null;
-  }
-
+export class Menu extends AbstractComponent {
   getTemplate() {
     return createSiteMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
