+++ src/components/site-sort.js
@@ -1,4 +1,5 @@
-import {createElement} from "../utils";
+
+import {AbstractComponent} from "./abstract-component.js";
 
 const createSiteSortTemplate = () => {
   return (
@@ -35,24 +36,8 @@ const createSiteSortTemplate = () => {
   );
 };
 
-export class Sorting {
-  constructor() {
-    this._element = null;
-  }
-
+export class Sorting extends AbstractComponent {
   getTemplate() {
     return createSiteSortTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
